done checking cities, huge upgrade is coming soon

+++ data/src/main/java/com/lab/esh1n/data/api/response/WindResponse.kt
@@ -8,5 +8,5 @@ data class WindResponse(
 	val deg: Int? = null,
 
 	@field:SerializedName("speed")
-	val speed: Int? = null
+	val speed: Float? = null
 )
